changed filenames, added UniformOrder, added CSS
made all 3 htm files so they have the same filetype (.htm not .html) and
same capitolization rules. Added the file UnifromOrder.htm for public
relations and uniform purhases and stuff like that -- it's more for me.
Also added some basic css styling for the webserver to make it look
nicer. Feel free to expand on this.